fix(schedule): fix schedule to not shift the next day, also add time/tzdata for time zone awareness
