+++ cmd/controller/main.go
@@ -11,6 +11,8 @@ import (
 	"os"
 	"time"
 
+	_ "time/tzdata"
+
 	"k8s.io/apimachinery/pkg/runtime"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/client-go/kubernetes"
