+++ internal/scheduler/schedule.go
@@ -248,8 +248,8 @@ func (e *ScheduleEvaluator) evaluateExtend(baseWindows, exceptionWindows []OffHo
 		return nil, fmt.Errorf("evaluate exception windows: %w", err)
 	}
 
-	// Union: hibernate if EITHER schedule says hibernate
-	shouldHibernate := baseResult.ShouldHibernate || exceptionResult.ShouldHibernate
+	// Hibernate when both schedule says hibernate
+	shouldHibernate := baseResult.ShouldHibernate && exceptionResult.ShouldHibernate
 
 	// Calculate next events (take the earlier of the two for each)
 	nextHibernate := earlierTime(baseResult.NextHibernateTime, exceptionResult.NextHibernateTime)
@@ -566,24 +566,13 @@ func ConvertOffHoursToCron(windows []OffHourWindow) (string, string, error) {
 	}
 
 	// Check if this is an overnight window (end time is before start time)
-	isOvernight := endHour < startHour || (endHour == startHour && endMin < startMin)
+	// This is used for logic validation if needed, but cron generation now uses same days
+	// isOvernight := endHour < startHour || (endHour == startHour && endMin < startMin)
 
 	// Build cron expressions
 	// Format: MIN HOUR DAY MONTH DOW
 	hibernateCron := fmt.Sprintf("%d %d * * %s", startMin, startHour, cronDays)
-
-	var wakeUpCron string
-	if isOvernight {
-		// For overnight windows, wake-up happens on the next day
-		wakeUpDays, err := shiftDaysForward(window.DaysOfWeek)
-		if err != nil {
-			return "", "", fmt.Errorf("failed to shift days for overnight window: %w", err)
-		}
-		wakeUpCron = fmt.Sprintf("%d %d * * %s", endMin, endHour, wakeUpDays)
-	} else {
-		// Same-day window
-		wakeUpCron = fmt.Sprintf("%d %d * * %s", endMin, endHour, cronDays)
-	}
+	wakeUpCron := fmt.Sprintf("%d %d * * %s", endMin, endHour, cronDays)
 
 	return hibernateCron, wakeUpCron, nil
 }
@@ -640,34 +629,4 @@ func convertDaysToCron(days []string) (string, error) {
 	}
 
 	return strings.Join(cronDays, ","), nil
-}
-
-// shiftDaysForward shifts each day forward by one (MON->TUE, SUN->MON).
-// Used for overnight windows where wake-up occurs on the next day.
-func shiftDaysForward(days []string) (string, error) {
-	dayOrder := []string{"SUN", "MON", "TUE", "WED", "THU", "FRI", "SAT"}
-	dayMap := map[string]int{
-		"SUN": 0,
-		"MON": 1,
-		"TUE": 2,
-		"WED": 3,
-		"THU": 4,
-		"FRI": 5,
-		"SAT": 6,
-	}
-
-	var shiftedDays []string
-	for _, day := range days {
-		dayUpper := strings.ToUpper(day)
-		idx, ok := dayMap[dayUpper]
-		if !ok {
-			return "", fmt.Errorf("invalid day %q", day)
-		}
-		// Shift forward by 1, wrapping around (SAT -> SUN)
-		nextIdx := (idx + 1) % 7
-		nextDay := dayMap[dayOrder[nextIdx]]
-		shiftedDays = append(shiftedDays, fmt.Sprintf("%d", nextDay))
-	}
-
-	return strings.Join(shiftedDays, ","), nil
 }
